requirements.txt
python-decouple
telethon
requests==2.32.3
python-telegram-bot